@@ -9,18 +9,18 @@
     },
     "dependencies": {
         "@types/angular-ui-bootstrap": "^1.0.0",
-        "@types/bootstrap": "^4.3.0",
-        "@types/jquery": "^3.3.29",
+        "@types/bootstrap": "^4.3.1",
+        "@types/jquery": "^3.3.31",
         "@types/notifyjs": "^3.0.0",
         "@types/select2": "^4.0.48",
         "@types/url-parse": "^1.4.3",
-        "angular": "^1.7.4",
+        "angular": "^1.7.8",
         "angular-inview": "^3.1.0",
         "angular1-star-rating": "^1.2.11",
-        "bitcoinjs-lib": "^4.0.3",
-        "bitcoinjs-message": "^2.0.0",
+        "bitcoinjs-lib": "^4.0.5",
+        "bitcoinjs-message": "^2.1.0",
         "bootstrap": "^4.3.1",
-        "buffer": "^5.2.1",
+        "buffer": "^5.4.3",
         "copy-to-clipboard": "^3.2.0",
         "giticon": "^1.0.8",
         "hashicon": "^0.3.0",
@@ -30,30 +30,30 @@
         "lodash": "^4.17.15",
         "notifyjs-browser": "^0.4.2",
         "popper.js": "^1.15.0",
-        "select2": "^4.0.8",
+        "select2": "^4.0.10",
         "ui-bootstrap4": "^3.0.6",
         "url-parse": "^1.4.7",
         "vis": "^4.21.0"
     },
     "devDependencies": {
-        "@types/angular": "^1.6.50",
+        "@types/angular": "^1.6.56",
         "@types/chrome": "0.0.73",
         "clean-webpack-plugin": "^0.1.19",
-        "copy-webpack-plugin": "^4.5.2",
-        "css-loader": "^1.0.0",
+        "copy-webpack-plugin": "^4.6.0",
+        "css-loader": "^1.0.1",
         "file-loader": "^2.0.0",
         "html-loader": "^0.5.5",
         "html-webpack-plugin": "3.0.7",
-        "rimraf": "^2.6.2",
-        "shelljs": "^0.8.2",
-        "style-loader": "^0.23.0",
-        "ts-loader": "^5.0.0",
-        "typescript": "^3.0.3",
+        "rimraf": "^2.7.1",
+        "shelljs": "^0.8.3",
+        "style-loader": "^0.23.1",
+        "ts-loader": "^5.4.5",
+        "typescript": "^3.6.3",
         "uglifyjs-webpack-plugin": "^1.3.0",
-        "url-loader": "^1.1.1",
-        "webextension-polyfill-ts": "^0.9.0",
-        "webpack": "^4.17.2",
-        "webpack-cli": "^3.1.0",
-        "webpack-dev-server": "^3.1.14"
+        "url-loader": "^1.1.2",
+        "webextension-polyfill-ts": "^0.9.1",
+        "webpack": "^4.40.2",
+        "webpack-cli": "^3.3.9",
+        "webpack-dev-server": "^3.8.1"
     }
 }
